[Tilemap] moved type definitions to their own files
