+++ Assets/Scripts/Tiles/ResourceTile.cs
@@ -1,0 +1,16 @@
+﻿namespace Untitled
+{
+    namespace Tiles
+    {
+        public class ResourceTile
+        {
+            public float Value { get; set; }
+            
+            public ResourceTile(float value)
+            {
+                Value = value;
+            }
+        }
+    }
+}
+
