+++ Assets/Scripts/Tiles/ResourceTile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5739c90a6e026304fbeac587647023b5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
