+++ Assets/Scripts/Tiles/TileType.cs
@@ -1,0 +1,14 @@
+﻿namespace Untitled
+{
+    namespace Tiles
+    {
+        public enum TileType
+        {
+            Coal,
+            Ground,
+            Other,
+            None
+        }
+    }
+}
+
