+++ Assets/Scripts/Tiles/TileType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b564260a69e4d6f4aa1e1416ed3881ec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
